delete testing-macros
