--- crates/testing-macros/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "hotshot-testing-macros"
-version = { workspace = true }
-edition = { workspace = true }
-description = "Macros for creating hotshot tests"
-
-[dependencies]
-ark-bls12-381 = { workspace = true }
-async-compatibility-layer = { workspace = true }
-async-trait = { workspace = true }
-# needed for vrf demo
-# so non-optional for now
-blake3 = { workspace = true }
-commit = { workspace = true }
-either = { workspace = true }
-futures = { workspace = true }
-hotshot = { path = "../hotshot", default-features = false }
-hotshot-types = { path = "../types" }
-hotshot-testing = { path = "../testing", default-features = false }
-hotshot-example-types = { path = "../example-types" }
-jf-primitives = { workspace = true }
-rand = { workspace = true }
-snafu = { workspace = true }
-tracing = { workspace = true }
-serde = { workspace = true }
-# proc macro stuff
-quote = "1"
-syn = { version = "2", features = ["full", "extra-traits"] }
-proc-macro2 = "1"
-derive_builder = "0.20"
-
-[dev-dependencies]
-async-lock = { workspace = true }
-
-[lib]
-proc-macro = true
-
-[target.'cfg(all(async_executor_impl = "tokio"))'.dependencies]
-tokio = { workspace = true }
-[target.'cfg(all(async_executor_impl = "async-std"))'.dependencies]
-async-std = { workspace = true }
-
-[lints]
-workspace = true
