+++ crates/testing/tests/cross_tests.rs
@@ -1,5 +1,6 @@
 use hotshot_example_types::node_types::{Libp2pImpl, MemoryImpl, PushCdnImpl, WebImpl};
 use hotshot_example_types::state_types::TestTypes;
+use hotshot_macros::cross_tests;
 use hotshot_testing::completion_task::{
     CompletionTaskDescription, TimeBasedCompletionTaskDescription,
 };
@@ -7,7 +8,6 @@ use hotshot_testing::spinning_task::ChangeNode;
 use hotshot_testing::spinning_task::SpinningTaskDescription;
 use hotshot_testing::spinning_task::UpDown;
 use hotshot_testing::test_builder::TestMetadata;
-use hotshot_testing_macros::cross_tests;
 use std::time::Duration;
 
 cross_tests!(
